21 refactor : 입력칸 크기조정, 이미지 업로드 hooks로 추출
